refactor fixtures
