+++ test/unit_tests/test_storage.py
@@ -1,48 +1,10 @@
-import os
-from pathlib import Path
 from typing import List, Tuple
 
-import boto3
 import pytest
 from botocore.exceptions import ClientError
-from moto import mock_s3
-from moto.core import DEFAULT_ACCOUNT_ID
 
-from moto.s3.models import S3Backend, s3_backends, FakeKey, FakeBucket
+from moto.s3.models import FakeKey, FakeBucket
 from werkzeug.datastructures import FileStorage
-
-from flaskr import CloudStorage
-
-
-@pytest.fixture
-def test_data_dir():
-    test_dir = Path(__file__).parent.parent
-    test_data_directory = test_dir.joinpath("test_data")
-    yield test_data_directory
-
-
-@pytest.fixture
-def aws_credentials():
-    """Mocked AWS Credentials for moto."""
-    os.environ["AWS_ACCESS_KEY_ID"] = "testing"
-    os.environ["AWS_SECRET_ACCESS_KEY"] = "testing"
-    os.environ["AWS_SECURITY_TOKEN"] = "testing"
-    os.environ["AWS_SESSION_TOKEN"] = "testing"
-
-
-@pytest.fixture
-def backend(aws_credentials):
-    s3_backend: S3Backend = s3_backends[DEFAULT_ACCOUNT_ID]["global"]
-    yield s3_backend
-
-
-@pytest.fixture
-def fake_storage(aws_credentials):
-    with mock_s3():
-        conn = boto3.client("s3", region_name="us-east-1")
-        conn.create_bucket(Bucket="jhb-crossword")
-
-        yield CloudStorage()
 
 
 def _pop_latest_item(fb: FakeBucket) -> Tuple[str, List[FakeKey]]:
@@ -111,19 +73,19 @@ def test_no_file_ext(fake_storage):
 
 
 @pytest.mark.parametrize("test_input", ["test.json"])
-def test_upload_puzzle_file(fake_storage, test_data_dir, backend, test_input):
+def test_upload_puzzle_file(fake_storage, test_data_dir, s3_backend, test_input):
     with open(test_data_dir.joinpath(test_input), "rb") as local_file:
         local_content = local_file.read()
     file = FileStorage(open(test_data_dir.joinpath(test_input), "rb"), name=test_input)
     fake_storage.upload_puzzle_json(file)
-    fb: FakeBucket = backend.get_bucket("jhb-crossword")
+    fb: FakeBucket = s3_backend.get_bucket("jhb-crossword")
     uploaded_content = pop_latest_item_content(fb)
     assert local_content == uploaded_content
 
 
 @pytest.mark.parametrize("test_input", ["test.json"])
 def test_upload_image_file_wrong_format(
-    fake_storage, test_data_dir, backend, test_input
+    fake_storage, test_data_dir, s3_backend, test_input
 ):
     with pytest.raises(ValueError) as excinfo:
         file = FileStorage(
@@ -135,7 +97,7 @@ def test_upload_image_file_wrong_format(
 
 @pytest.mark.parametrize("test_input", ["test.png"])
 def test_upload_json_file_wrong_format(
-    fake_storage, test_data_dir, backend, test_input
+    fake_storage, test_data_dir, s3_backend, test_input
 ):
     with pytest.raises(ValueError) as excinfo:
         file = FileStorage(
@@ -168,32 +130,32 @@ def test_invalid_json_keys_upload(fake_storage, test_data_dir, test_input):
 
 
 @pytest.mark.parametrize("test_input", ["test.png"])
-def test_upload_image(fake_storage, test_data_dir, backend, test_input):
+def test_upload_image(fake_storage, test_data_dir, s3_backend, test_input):
     with open(test_data_dir.joinpath(test_input), "rb") as local_file:
         local_content = local_file.read()
     file = FileStorage(open(test_data_dir.joinpath(test_input), "rb"), name=test_input)
     fake_storage.upload_image(file)
-    fb: FakeBucket = backend.get_bucket("jhb-crossword")
+    fb: FakeBucket = s3_backend.get_bucket("jhb-crossword")
     uploaded_content = pop_latest_item_content(fb)
     assert local_content == uploaded_content
 
 
-def test_download_image(fake_storage, backend, test_data_dir):
+def test_download_image(fake_storage, s3_backend, test_data_dir):
     fname = "test.png"
     with open(test_data_dir.joinpath(fname), "rb") as file:
         uploaded_bytes = file.read()
         file.seek(0)
-        backend.put_object("jhb-crossword", key_name=fname, value=file.read())
+        s3_backend.put_object("jhb-crossword", key_name=fname, value=file.read())
     downloaded_bytes = fake_storage.download_image(fname)
     assert downloaded_bytes.read() == uploaded_bytes
 
 
-def test_download_json(fake_storage, backend, test_data_dir):
+def test_download_json(fake_storage, s3_backend, test_data_dir):
     fname = "test.json"
     with open(test_data_dir.joinpath(fname), "rb") as file:
         uploaded_bytes = file.read()
         file.seek(0)
-        backend.put_object("jhb-crossword", key_name=fname, value=file.read())
+        s3_backend.put_object("jhb-crossword", key_name=fname, value=file.read())
     downloaded_bytes = fake_storage.download_puzzle_json(fname)
     assert downloaded_bytes.read() == uploaded_bytes
