+++ test/unit_tests/test_user_databases.py
@@ -1,101 +1,5 @@
 #  TODO: are these integration tests?
-import os
-
-import boto3
 import pytest
-from moto import mock_dynamodb
-from moto.dynamodb.models import dynamodb_backends, DynamoDBBackend
-from moto.core import DEFAULT_ACCOUNT_ID
-
-from flaskr import UserDatabase
-
-
-@pytest.fixture
-def db_backend(aws_credentials):
-    db_backend: DynamoDBBackend = dynamodb_backends[DEFAULT_ACCOUNT_ID]["eu-north-1"]
-    yield db_backend
-
-
-@pytest.fixture
-def aws_credentials():
-    """Mocked AWS Credentials for moto."""
-    os.environ["AWS_ACCESS_KEY_ID"] = "testing"
-    os.environ["AWS_SECRET_ACCESS_KEY"] = "testing"
-    os.environ["AWS_SECURITY_TOKEN"] = "testing"
-    os.environ["AWS_SESSION_TOKEN"] = "testing"
-
-
-@pytest.fixture
-def fake_crossword_user_databases(aws_credentials):
-    with mock_dynamodb():
-        user_data_table_params = {
-            "TableName": "crossword-userdata",
-            "KeySchema": [
-                {"AttributeName": "id", "KeyType": "HASH"},
-            ],
-            "AttributeDefinitions": [
-                {"AttributeName": "id", "AttributeType": "S"},
-            ],
-            "ProvisionedThroughput": {
-                "ReadCapacityUnits": 10,
-                "WriteCapacityUnits": 10,
-            },
-        }
-        user_email_table_params = {
-            "TableName": "crossword-emails",
-            "KeySchema": [
-                {"AttributeName": "email", "KeyType": "HASH"},
-            ],
-            "AttributeDefinitions": [
-                {"AttributeName": "email", "AttributeType": "S"},
-            ],
-            "ProvisionedThroughput": {
-                "ReadCapacityUnits": 10,
-                "WriteCapacityUnits": 10,
-            },
-        }
-        user_name_table_params = {
-            "TableName": "crossword-usernames",
-            "KeySchema": [
-                {"AttributeName": "username", "KeyType": "HASH"},
-            ],
-            "AttributeDefinitions": [
-                {"AttributeName": "username", "AttributeType": "S"},
-            ],
-            "ProvisionedThroughput": {
-                "ReadCapacityUnits": 10,
-                "WriteCapacityUnits": 10,
-            },
-        }
-
-        conn = boto3.client("dynamodb")
-        conn.create_table(**user_data_table_params)
-        conn.create_table(**user_email_table_params)
-        conn.create_table(**user_name_table_params)
-
-        yield UserDatabase()
-
-
-@pytest.fixture
-def user_test_data():
-    userdata = {
-        "id": {"S": "test"},
-        "username": {"S": "uname"},
-        "password": {"S": "pword"},
-        "email": {"S": "email@addr.com"},
-        "resetGuid": {"S": "123"},
-    }
-
-    username = {
-        "username": {"S": "uname"},
-        "id": {"S": "test"},
-    }
-    email = {
-        "email": {"S": "email@addr.com"},
-        "id": {"S": "test"},
-    }
-
-    return userdata, username, email
 
 
 class TestUserDatabase:
@@ -110,7 +14,7 @@ class TestUserDatabase:
         ],
     )
     def test_add_user_main_database(
-        self, fake_crossword_user_databases, db_backend, key, val
+        self, fake_crossword_user_databases, user_db_backend, key, val
     ):
         fake_crossword_user_databases.register_new_user(
             user_id="test",
@@ -119,10 +23,12 @@ class TestUserDatabase:
             email="email@addr.com",
             reset_guid="123",
         )
-        data = db_backend.get_item("crossword-userdata", {"id": {"S": "test"}})
+        data = user_db_backend.get_item("crossword-userdata", {"id": {"S": "test"}})
         assert data.attrs[key].value == val
 
-    def test_add_user_email_database(self, fake_crossword_user_databases, db_backend):
+    def test_add_user_email_database(
+        self, fake_crossword_user_databases, user_db_backend
+    ):
         fake_crossword_user_databases.register_new_user(
             user_id="test",
             username="uname",
@@ -130,13 +36,13 @@ class TestUserDatabase:
             email="email@addr.com",
             reset_guid="123",
         )
-        data = db_backend.get_item(
+        data = user_db_backend.get_item(
             "crossword-emails", {"email": {"S": "email@addr.com"}}
         )
         assert data.attrs["id"].value == "test"
 
     def test_add_user_username_database(
-        self, fake_crossword_user_databases, db_backend
+        self, fake_crossword_user_databases, user_db_backend
     ):
         fake_crossword_user_databases.register_new_user(
             user_id="test",
@@ -145,10 +51,12 @@ class TestUserDatabase:
             email="email@addr.com",
             reset_guid="123",
         )
-        data = db_backend.get_item("crossword-usernames", {"username": {"S": "uname"}})
+        data = user_db_backend.get_item(
+            "crossword-usernames", {"username": {"S": "uname"}}
+        )
         assert data.attrs["id"].value == "test"
 
-    def test_get_user_by_id(self, fake_crossword_user_databases, db_backend):
+    def test_get_user_by_id(self, fake_crossword_user_databases, user_db_backend):
         fake_crossword_user_databases.register_new_user(
             user_id="test",
             username="uname",
@@ -159,7 +67,7 @@ class TestUserDatabase:
         response = fake_crossword_user_databases.get_id_for_username("uname")
         assert response["Item"]["id"] == "test"
 
-    def test_get_id_for_email(self, fake_crossword_user_databases, db_backend):
+    def test_get_id_for_email(self, fake_crossword_user_databases, user_db_backend):
         fake_crossword_user_databases.register_new_user(
             user_id="test",
             username="uname",
@@ -180,7 +88,9 @@ class TestUserDatabase:
             ("resetGuid", "123"),
         ],
     )
-    def test_get_user_data(self, fake_crossword_user_databases, db_backend, key, val):
+    def test_get_user_data(
+        self, fake_crossword_user_databases, user_db_backend, key, val
+    ):
         fake_crossword_user_databases.register_new_user(
             user_id="test",
             username="uname",
@@ -212,7 +122,7 @@ class TestUserDatabase:
         assert not registered_second_user
 
     def test_add_duplicate_username_backend(
-        self, fake_crossword_user_databases, db_backend
+        self, fake_crossword_user_databases, user_db_backend
     ):
         fake_crossword_user_databases.register_new_user(
             user_id="expected_id",
@@ -230,10 +140,12 @@ class TestUserDatabase:
             reset_guid="123",
         )
 
-        data = db_backend.get_item("crossword-usernames", {"username": {"S": "uname"}})
+        data = user_db_backend.get_item(
+            "crossword-usernames", {"username": {"S": "uname"}}
+        )
         assert data.attrs["id"].value == "expected_id"
 
-    def test_add_duplicate_email(self, fake_crossword_user_databases, db_backend):
+    def test_add_duplicate_email(self, fake_crossword_user_databases, user_db_backend):
         fake_crossword_user_databases.register_new_user(
             user_id="expected_id",
             username="uname",
@@ -250,12 +162,12 @@ class TestUserDatabase:
             reset_guid="123",
         )
 
-        data = db_backend.get_item(
+        data = user_db_backend.get_item(
             "crossword-emails", {"email": {"S": "email@addr.com"}}
         )
         assert data.attrs["id"].value == "expected_id"
 
-    def test_delete_email(self, fake_crossword_user_databases, db_backend):
+    def test_delete_email(self, fake_crossword_user_databases, user_db_backend):
         fake_crossword_user_databases.register_new_user(
             user_id="expected_id",
             username="uname",
@@ -266,7 +178,7 @@ class TestUserDatabase:
 
         fake_crossword_user_databases.delete_email("email@addr.com")
 
-        data = db_backend.get_item(
+        data = user_db_backend.get_item(
             "crossword-emails", {"email": {"S": "email@addr.com"}}
         )
         assert data is None
@@ -281,7 +193,9 @@ class TestUserDatabase:
             ("resetGuid", "456"),
         ],
     )
-    def test_update_password(self, fake_crossword_user_databases, db_backend, key, val):
+    def test_update_password(
+        self, fake_crossword_user_databases, user_db_backend, key, val
+    ):
         fake_crossword_user_databases.register_new_user(
             user_id="test",
             username="uname",
@@ -291,5 +205,5 @@ class TestUserDatabase:
         )
 
         fake_crossword_user_databases.update_password("test", "pword2", "456")
-        data = db_backend.get_item("crossword-userdata", {"id": {"S": "test"}})
+        data = user_db_backend.get_item("crossword-userdata", {"id": {"S": "test"}})
         assert data.attrs[key].value == val
